+++ packages/optimistic-oracle/index.js
@@ -3,7 +3,14 @@
 require("dotenv").config();
 const retry = require("async-retry");
 
-const { Logger, waitForLogger, delay, OptimisticOracleClient, GasEstimator } = require("@uma/financial-templates-lib");
+const {
+  Logger,
+  waitForLogger,
+  delay,
+  OptimisticOracleClient,
+  GasEstimator,
+  OptimisticOracleType,
+} = require("@uma/financial-templates-lib");
 const { OptimisticOracleProposer } = require("./src/proposer");
 
 // Contract ABIs and network Addresses.
@@ -32,6 +39,7 @@ const OracleType = {
  * @param {Object} [commonPriceFeedConfig] Common configuration to pass to all PriceFeeds constructed by proposer.
  * @param {Object} [optimisticOracleProposerConfig] Configuration to construct the OptimisticOracle proposer.
  * @param {OracleType} [oracleType] Type of "Oracle" for this network, defaults to "Voting"
+ * @param {OptimisticOracleType} [optimisticOracleType] Type of "OptimisticOracle" for this network, defaults to "OptimisticOracle"
  * @return None or throws an Error.
  */
 async function run({
@@ -43,10 +51,11 @@ async function run({
   commonPriceFeedConfig,
   optimisticOracleProposerConfig,
   oracleType = OracleType.Voting,
+  optimisticOracleType = OptimisticOracleType.OptimisticOracle,
 }) {
   try {
     const [accounts, networkId] = await Promise.all([web3.eth.getAccounts(), web3.eth.net.getId()]);
-    const optimisticOracleAddress = await getAddress("OptimisticOracle", networkId);
+    const optimisticOracleAddress = await getAddress(optimisticOracleType, networkId);
     // If pollingDelay === 0 then the bot is running in serverless mode and should send a `debug` level log.
     // Else, if running in loop mode (pollingDelay != 0), then it should send a `info` level log.
     logger[pollingDelay === 0 ? "debug" : "info"]({
@@ -59,17 +68,20 @@ async function run({
       commonPriceFeedConfig,
       optimisticOracleProposerConfig,
       oracleType,
+      optimisticOracleType,
     });
 
     // Create the OptimisticOracleClient to query on-chain information, GasEstimator to get latest gas prices and an
     // instance of the OO Proposer to respond to price requests and proposals.
     const optimisticOracleClient = new OptimisticOracleClient(
       logger,
-      getAbi("OptimisticOracle"),
+      getAbi(optimisticOracleType),
       getAbi(oracleType),
       web3,
       optimisticOracleAddress,
-      await getAddress(oracleType, networkId)
+      await getAddress(oracleType, networkId),
+      604800, // default lookback setting for this client
+      optimisticOracleType
     );
     const gasEstimator = new GasEstimator(logger, 60, networkId);
 
@@ -159,6 +171,11 @@ async function Poll(callback) {
       // Type of "Oracle" to that optimistic oracle on this network submits price requests to, default is "Voting".
       // The other possible types are exported in an enum in this file.
       oracleType: process.env.ORACLE_TYPE ? process.env.ORACLE_TYPE : OracleType.Voting,
+      // Type of "OptimisticOracle" to load in client, default is OptimisticOracle. The other possible types are
+      // exported in an enum from financial-templates-lib/OptimisticOracleClient.
+      optimisticOracleType: process.env.OPTIMISTIC_ORACLE_TYPE
+        ? process.env.OPTIMISTIC_ORACLE_TYPE
+        : OptimisticOracleType.OptimisticOracle,
     };
 
     await run({ logger: Logger, web3: getWeb3(), ...executionParameters });