+++ packages/api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.3](https://github.com/UMAprotocol/protocol/compare/@uma/api@0.8.2...@uma/api@0.8.3) (2021-11-05)
+
+**Note:** Version bump only for package @uma/api
+
 ## [0.8.2](https://github.com/UMAprotocol/protocol/compare/@uma/api@0.8.1...@uma/api@0.8.2) (2021-11-02)
 
 **Note:** Version bump only for package @uma/api