+++ packages/fx-tunnel-relayer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.9](https://github.com/UMAprotocol/protocol/compare/@uma/fx-tunnel-relayer@1.0.8...@uma/fx-tunnel-relayer@1.0.9) (2021-11-05)
+
+**Note:** Version bump only for package @uma/fx-tunnel-relayer
+
 ## [1.0.8](https://github.com/UMAprotocol/protocol/compare/@uma/fx-tunnel-relayer@1.0.7...@uma/fx-tunnel-relayer@1.0.8) (2021-11-02)
 
 **Note:** Version bump only for package @uma/fx-tunnel-relayer